requirements.txt
yfinance
psycopg2-binary
groq
python-dotenv
fastapi
uvicorn
sqlalchemy
pandas
